+++ api/src/main/java/org/allaymc/api/eventbus/event/server/PlayerJoinEvent.java
@@ -2,9 +2,11 @@ package org.allaymc.api.eventbus.event.server;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 import org.allaymc.api.annotation.CallerThread;
 import org.allaymc.api.annotation.ThreadType;
 import org.allaymc.api.eventbus.event.CancellableEvent;
+import org.allaymc.api.message.MayContainTrKey;
 import org.allaymc.api.player.Player;
 
 /**
@@ -14,10 +16,25 @@ import org.allaymc.api.player.Player;
  */
 @CallerThread(ThreadType.NETWORK)
 @AllArgsConstructor
+@Getter
 public class PlayerJoinEvent extends ServerEvent implements CancellableEvent {
+
     /**
      * The player associated with this event.
      */
-    @Getter
     protected Player player;
+
+    /**
+     * The message that will be broadcast if the event is not cancelled.
+     */
+    @MayContainTrKey
+    @Setter
+    protected String joinMessage;
+
+    /**
+     * The reason that will be shown to the player if the event is cancelled.
+     */
+    @MayContainTrKey
+    @Setter
+    protected String disconnectReason;
 }